+++ Proyecto Backend/src/controllers/productsController.js
@@ -1,4 +1,5 @@
-import { Product } from "../models/productoModel.js";
+import { Product } from "../models/productModel.js";
+import mongoose from "mongoose";
 
 export const createProduct = async (req, res, next) => {
   try {
@@ -8,7 +9,7 @@ export const createProduct = async (req, res, next) => {
       const error = new Error(
         "Todos los campos son obligatorios: name, category, price y stock."
       );
-      error.status = 400;
+      error.statusCode = 400;
       return next(error);
     }
 
@@ -23,14 +24,14 @@ export const createProduct = async (req, res, next) => {
 
 export const getProducts = async (req, res, next) => {
   try {
-    const products = await Product.find();
+    const products = await Product.find().sort({ createdAt: 1 });
     res.status(200).json(products);
   } catch (error) {
     next(error);
   }
 };
 
-export const getProductBy = async (req, res, next) => {
+export const findProduct = async (req, res, next) => {
   try {
     const { category, name } = req.query;
     const filter = {};
@@ -49,6 +50,12 @@ export const updateProduct = async (req, res, next) => {
     const { id } = req.params;
     const { name, category, price, stock } = req.body;
 
+    if (!id || !mongoose.Types.ObjectId.isValid(id)) {
+      const err = new Error("ID inválido o no proporcionado.");
+      err.statusCode = 400;
+      throw err;
+    }
+
     const updatedProduct = await Product.findByIdAndUpdate(
       id,
       { name, category, price, stock },
@@ -57,7 +64,7 @@ export const updateProduct = async (req, res, next) => {
 
     if (!updatedProduct) {
       const error = new Error("Producto no encontrado");
-      error.status = 404;
+      error.statusCode = 404;
       return next(error);
     }
 
@@ -67,15 +74,21 @@ export const updateProduct = async (req, res, next) => {
   }
 };
 
-
 export const deleteProduct = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!id || !mongoose.Types.ObjectId.isValid(id)) {
+      const error = new Error("ID inválido o no proporcionado.");
+      error.statusCode = 400;
+      throw error;
+    }
+
     const deletedProduct = await Product.findByIdAndDelete(id);
 
     if (!deletedProduct) {
       const error = new Error("Producto no encontrado");
-      error.status = 404;
+      error.statusCode = 404;
       return next(error);
     }
